backend/requirements.txt
fastapi
uvicorn
google-cloud-vision
python-multipart
supabase
python-dotenv
fastapi>=0.110
uvicorn>=0.30
pillow>=9.0.0
numpy>=1.21.0
psycopg[binary]>=3.1.0
open-clip-torch>=2.20.0
torch>=2.0.0
torchvision>=0.15.0
python-dotenv>=1.0.1
requests>=2.31.0
